[refactor] :recycle: refactor mint tokens
